+++ packages/codegen/src/generate-descriptors.ts
@@ -301,10 +301,6 @@ type AnonymousEnum<T extends {}> = T & {
   __anonymous: true
 }
 
-type IEnum<T extends {}> = Enum<{
-  [K in keyof T & string]: { type: K, value: T[K] }
-}[keyof T & string]>
-
 type MyTuple<T> = [T, ...T[]]
 
 type SeparateUndefined<T> = undefined extends T
@@ -324,10 +320,10 @@ type Anonymize<T> = SeparateUndefined<
     | symbol
     | Binary
     | Uint8Array
-    | Enum<{ type: string; value: any }>
+    | Enum<any>
     ? T
     : T extends AnonymousEnum<infer V>
-      ? IEnum<V>
+      ? Enum<V>
       : T extends MyTuple<any>
         ? {
             [K in keyof T]: T[K]