+++ packages/codegen/src/generate-types.ts
@@ -32,10 +32,6 @@ export const generateTypes = (
     __anonymous: true
   }
   
-  type IEnum<T extends {}> = Enum<{
-    [K in keyof T & string]: { type: K, value: T[K] }
-  }[keyof T & string]>
-  
   type MyTuple<T> = [T, ...T[]]
   
   type SeparateUndefined<T> = undefined extends T
@@ -55,10 +51,10 @@ export const generateTypes = (
       | symbol
       | Binary
       | Uint8Array
-      | Enum<{ type: string; value: any }>
+      | Enum<any>
       ? T
       : T extends AnonymousEnum<infer V>
-        ? IEnum<V>
+        ? Enum<V>
         : T extends MyTuple<any>
           ? {
               [K in keyof T]: T[K]