+++ packages/client/CHANGELOG.md
@@ -5,6 +5,12 @@
 ### Added
 
 - `getChainSpecData: () => Promise<{name: string, genesisHash: string, properties: any}>`.
+- New type `EnumVariant<T, K>` to select one specific variant from an enum.
+- Improved Enum type inference so that they can be assigned between types as long as they are compatible.
+
+### Breaking
+
+- The generic for `Enum<T>` is now an object of `{ [type: string]: any }`, rather than a union of `{ type: string, value: any } | ... | { type: string, value: any }`
 
 ## 0.4.0 - 2024-04-23