+++ packages/client/CHANGELOG.md
@@ -2,15 +2,22 @@
 
 ## Unreleased
 
+## 0.5.0 - 2024-04-25
+
+### Breaking
+
+- The generic for `Enum<T>` is now an object of `{ [type: string]: any }`, rather than a union of `{ type: string, value: any } | ... | { type: string, value: any }`
+
 ### Added
 
 - `getChainSpecData: () => Promise<{name: string, genesisHash: string, properties: any}>`.
 - New type `EnumVariant<T, K>` to select one specific variant from an enum.
 - Improved Enum type inference so that they can be assigned between types as long as they are compatible.
 
-### Breaking
+### Fixed
 
-- The generic for `Enum<T>` is now an object of `{ [type: string]: any }`, rather than a union of `{ type: string, value: any } | ... | { type: string, value: any }`
+- substrate-client: Improve the compatibility handling with the different versions of the JSON-RPC methods.
+- Upgraded to `smoldot@2.0.24`
 
 ## 0.4.0 - 2024-04-23